[refactor/#100] getRelatedSearchWords 조회 로직을 SearchBar.tsx에서 AutoSearchList.tsx로 이동
연관 검색어 API 호출 로직(getRelatedSearchWords)을 SearchBar.tsx에서
AutoSearchList.tsx로 이동.
